pyproject.toml
[tool.poetry]
name = "nonebot-plugin-follow-withdraw"
version = "0.1.0"
description = "A plugin for nonebot2 that follows the user's withdraw of a command."
authors = ["CMHopeSunshine <277073121@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_follow_withdraw"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-plugin-datastore = "^1.0.0"

[tool.poetry.group.dev.dependencies]

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pycln]
path = "."

[tool.nonebot]
plugins = ["nonebot_plugin_follow_withdraw"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
